==> src/04_reg_analysis.Rmd <==
---
title: "Regression analysis"
author: "Jae Yeon Kim"
output:
  html_document: 
    toc: true
    theme: united
---

#. Setup 

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(
        tidyverse, # for the tidyverse framework 
        purrr, # for functional programming
        glue, # putting strings and objects together
        patchwork, # putting ggplots together 
        sjPlot, # for visualizing regressions
        ggrepel, 
        conflicted, # for resolving conflicting functions
        ggthemes, # for fancy ggplot themes
        here, # for self-contained projects
        ggpubr, # for pub-ready themes
        estimatr # stat modeling
)

# Prefer select from dplyr 
conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")

# for publication-friendly theme 
theme_set(theme_pubr(base_size = 10))

source(here("functions", "utils.r"))
```

# Importing files 

```{r}
df <- read.csv(here("processed_data", "augmented_df.csv"))#[,-1]

names(df)
```

# Regression analysis

## Create a nested dataframe 

```{r}
nested <- df %>%
  group_by(race) %>%
  nest()
```

## Run regression models 

Run model based on different dependent variables.

```{r}
# --- UNWEIGHTED MODELS (Robustness) ---

# Linked fate
lf_models_unweighted <- nested %>%
  mutate(
    race_mls = map(data, ~ols(., "linkedfate_race")),
    ethnic_mls = map(data, ~ols(., "linkedfate_ethnic"))
    )

comm_models_unweighted <- nested %>%
  mutate( 
    commrace = map(data, ~ols(., "commrace")),
    commcult = map(data, ~ols(.,"commcult")),
    commecon = map(data, ~ols(., "commecon")),      commpol = map(data, ~ols(.,"commpol"))
  )

fed_models_unweighted <- nested %>%
    mutate(
        inequality = map(data, ~ols(., "inequality")),
        bank_reg = map(data, ~ols(., "bank_reg")),
        minimum_wage = map(data, ~ols(., "minimum_wage")),
        rich_tax = map(data, ~ols(., "rich_tax")),
        free_college = map(data, ~ols(., "free_college"))
    )
    
# --- WEIGHTED MODELS (Main Analysis) ---

# Linked fate
lf_models_weighted <- nested %>%
  mutate(
    race_mls = map(data, ~ols(., "linkedfate_race", weights = "weight")),
    ethnic_mls = map(data, ~ols(., "linkedfate_ethnic", weights = "weight"))
    )

comm_models_weighted <- nested %>%
  mutate( 
    commrace = map(data, ~ols(., "commrace", weights = "weight")),
    commcult = map(data, ~ols(.,"commcult", weights = "weight")),
    commecon = map(data, ~ols(., "commecon", weights = "weight")),
    commpol = map(data, ~ols(.,"commpol", weights = "weight"))
  )

fed_models_weighted <- nested %>%
    mutate(
        inequality = map(data, ~ols(., "inequality", weights = "weight")),
        bank_reg = map(data, ~ols(., "bank_reg", weights = "weight")),
        minimum_wage = map(data, ~ols(., "minimum_wage", weights = "weight")),
        rich_tax = map(data, ~ols(., "rich_tax", weights = "weight")),
        free_college = map(data, ~ols(., "free_college", weights = "weight"))
    )
```

## Bind results 

```{r}
# --- Bind UNWEIGHTED Results ---
lf_res_unweighted <- reduce(list(unnest_model(lf_models_unweighted, "race_mls", "Race"),
unnest_model(lf_models_unweighted %>% filter(race %in% c("AAPI", "Latino")), "ethnic_mls", "Ethnic")), full_join)

comm_res_unweighted <- reduce(list(
  unnest_model(comm_models_unweighted, "commrace", "Racial"),
  unnest_model(comm_models_unweighted, "commcult", "Cultural"), 
  unnest_model(comm_models_unweighted, "commecon", "Economic"),
  unnest_model(comm_models_unweighted, "commpol", "Political")
), full_join)

fed_res_unweighted <- reduce(list(
  unnest_model(fed_models_unweighted, "inequality", "Reducing inequality"),
  unnest_model(fed_models_unweighted, "bank_reg", "Regulating banks"),
  unnest_model(fed_models_unweighted, "minimum_wage", "Minimum wage"),
  unnest_model(fed_models_unweighted, "rich_tax", "Taxing Rich"),
  unnest_model(fed_models_unweighted, "free_college", "Free college")
), full_join)

# --- Bind WEIGHTED Results ---
lf_res_weighted <- reduce(list(unnest_model(lf_models_weighted, "race_mls", "Race"),
unnest_model(lf_models_weighted %>% filter(race %in% c("AAPI", "Latino")), "ethnic_mls", "Ethnic")), full_join)

comm_res_weighted <- reduce(list(
  unnest_model(comm_models_weighted, "commrace", "Racial"),
  unnest_model(comm_models_weighted, "commcult", "Cultural"), 
  unnest_model(comm_models_weighted, "commecon", "Economic"),
  unnest_model(comm_models_weighted, "commpol", "Political")
), full_join)

fed_res_weighted <- reduce(list(
  unnest_model(fed_models_weighted, "inequality", "Reducing inequality"),
  unnest_model(fed_models_weighted, "bank_reg", "Regulating banks"),
  unnest_model(fed_models_weighted, "minimum_wage", "Minimum wage"),
  unnest_model(fed_models_weighted, "rich_tax", "Taxing Rich"),
  unnest_model(fed_models_weighted, "free_college", "Free college")
), full_join)
```

## Visualize outputs 

```{r}
# --- Visualize UNWEIGHTED ---
lf_res_unweighted %>%
  viz_ols("Linked fate (Unweighted)")
ggsave(here("outputs", "lf_res_unweighted.png"), width = 8)

comm_res_unweighted %>%
  viz_ols("Common factors (Unweighted)")
ggsave(here("outputs", "comm_res_unweighted.png"), width = 8, height = 8)

fed_res_unweighted %>%
  viz_ols("Federal governmenet policies (Unweighted)")
ggsave(here("outputs", "fed_res_unweighted.png"), width = 8, height = 8)

# --- Visualize WEIGHTED ---
lf_res_weighted %>%
  viz_ols("Linked fate (Weighted)")
ggsave(here("outputs", "lf_res_weighted.png"), width = 8)

comm_res_weighted %>%
  viz_ols("Common factors (Weighted)")
ggsave(here("outputs", "comm_res_weighted.png"), width = 8, height = 8)

fed_res_weighted %>%
  viz_ols("Federal governmenet policies (Weighted)")
ggsave(here("outputs", "fed_res_weighted.png"), width = 8, height = 8)
```
